This is the first version with the functionality that allows a professor to share content to the entire course without allowing it outside of the cohorts...It seems to be working but I want to make a number of changes to make it more powerful.
For instance, would it be good for a student to have this capability?  I'm thinking maybe yes.  For instance, when a student creates content, they can keep it private, share just within the cohort, share to all cohorts within the course or share completely public.  This seems like a good idea...but I need to think it through more.  

This seems like a good point in time to learn how to branch or fork a git repository.